refactored and made AddChatbotDocumentModal its own component
