+++ packages/frontend/app/(dashboard)/course/[cid]/(settings)/settings/chatbot_settings/page.tsx
@@ -1,26 +1,7 @@
 'use client'
 
-import {
-  Button,
-  Form,
-  Input,
-  Modal,
-  Pagination,
-  Progress,
-  Switch,
-  Table,
-  Tooltip,
-  message,
-} from 'antd'
+import { Button, Input, Pagination, Progress, Table, message } from 'antd'
 import { ReactElement, useCallback, useEffect, useState } from 'react'
-import {
-  FileAddOutlined,
-  GithubOutlined,
-  InboxOutlined,
-  QuestionCircleOutlined,
-} from '@ant-design/icons'
-import { RcFile } from 'antd/lib/upload'
-import Dragger from 'antd/lib/upload/Dragger'
 import axios from 'axios'
 import { useUserInfo } from '@/app/contexts/userContext'
 import { SourceDocument } from '../chatbot_questions/page'
@@ -28,6 +9,7 @@ import Link from 'next/link'
 import { TableRowSelection } from 'antd/es/table/interface'
 import ChatbotSettingsModal from './components/ChatbotSettingsModal'
 import Highlighter from 'react-highlight-words'
+import AddChatbotDocumentModal from './components/AddChatbotDocumentModal'
 
 export interface ChatbotDocument {
   id: number
@@ -48,16 +30,11 @@ export default function ChatbotSettings({
   params,
 }: ChatbotPanelProps): ReactElement {
   const courseId = Number(params.cid)
-  const [form] = Form.useForm()
-  const { userInfo } = useUserInfo()
   const [chatbotParameterModalOpen, setChatbotParameterModalOpen] =
     useState(false)
   const [addDocumentModalOpen, setAddDocumentModalOpen] = useState(false)
-  const [isSlideDeck, setIsSlideDeck] = useState(false)
-  const [documentType, setDocumentType] = useState('FILE')
+  const { userInfo } = useUserInfo()
   const [search, setSearch] = useState('')
-  const [loading, setLoading] = useState(false)
-  const [countProcessed, setCountProcessed] = useState(0)
   const [selectViewEnabled, setSelectViewEnabled] = useState(false)
   const [selectedRowKeys, setSelectedRowKeys] = useState<React.Key[]>([])
   const [totalDocuments, setTotalDocuments] = useState(0)
@@ -65,19 +42,8 @@ export default function ChatbotSettings({
   const [filteredDocuments, setFilteredDocuments] = useState<SourceDocument[]>(
     [],
   )
-
-  const [fileList, setFileList] = useState<any[]>([])
-
-  const props = {
-    name: 'file',
-    multiple: true,
-    accept: '.docx,.pptx,.txt,.csv,.pdf',
-    fileList,
-    onChange(info: any) {
-      setFileList(info.fileList)
-    },
-    beforeUpload: () => false, // Prevent automatic upload
-  }
+  const [loading, setLoading] = useState(false)
+  const [countProcessed, setCountProcessed] = useState(0)
 
   const rowSelection: TableRowSelection<SourceDocument> = {
     type: 'checkbox',
@@ -88,8 +54,6 @@ export default function ChatbotSettings({
   const hasSelected = selectedRowKeys.length > 0
 
   const handleDeleteSelectedDocuments = async () => {
-    setLoading(true)
-    setCountProcessed(0)
     try {
       for (const docId of selectedRowKeys) {
         await fetch(`/chat/${courseId}/${docId}/document`, {
@@ -99,7 +63,7 @@ export default function ChatbotSettings({
             HMS_API_TOKEN: userInfo.chat_token.token,
           },
         })
-        setCountProcessed((prev) => prev + 1)
+        setCountProcessed(countProcessed + 1)
       }
       message.success('Documents deleted.')
       getDocuments()
@@ -108,6 +72,34 @@ export default function ChatbotSettings({
     } finally {
       setSelectViewEnabled(false)
       setSelectedRowKeys([])
+      setLoading(false)
+      setCountProcessed(0)
+    }
+  }
+
+  const handleDeleteDocument = async (record: any) => {
+    setLoading(true)
+    try {
+      const response = await fetch(
+        `/chat/${courseId}/${record.docId}/document`,
+        {
+          method: 'DELETE',
+          headers: {
+            'Content-Type': 'application/json',
+            HMS_API_TOKEN: userInfo.chat_token.token,
+          },
+        },
+      )
+
+      if (!response.ok) {
+        throw new Error(`Failed to upload ${File.name}`)
+      }
+
+      message.success('Document deleted.')
+      getDocuments()
+    } catch (e) {
+      message.error('Failed to delete document.')
+    } finally {
       setLoading(false)
     }
   }
@@ -199,15 +191,12 @@ export default function ChatbotSettings({
       console.error(e)
       setChatbotDocuments([])
       setTotalDocuments(0)
-    } finally {
-      setLoading(false)
     }
   }, [
     courseId,
     userInfo.chat_token.token,
     setChatbotDocuments,
     setTotalDocuments,
-    setLoading,
   ])
 
   useEffect(() => {
@@ -221,267 +210,14 @@ export default function ChatbotSettings({
     setFilteredDocuments(filtered)
   }, [search, chatbotDocuments])
 
-  const addUrl = async (url: string) => {
-    setLoading(true)
-    try {
-      const data = {
-        url: url,
-      }
-
-      const response = await fetch(`/chat/${courseId}/document/url/github`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          HMS_API_TOKEN: userInfo.chat_token.token,
-        },
-        body: JSON.stringify(data),
-      })
-
-      if (response.ok) {
-        message.success('File uploaded.')
-        getDocuments()
-      } else {
-        message.warning(
-          `Failed to upload file, please check the file type of linked document`,
-        )
-      }
-    } catch (e) {
-      console.log(e)
-    } finally {
-      setLoading(false)
-    }
-  }
-
-  const uploadFiles = async (files: RcFile[], source: string) => {
-    setCountProcessed(0)
-    for (const file of files) {
-      try {
-        const formData = new FormData()
-        formData.append('file', file)
-        // Create a JSON object and convert it to a string
-        const jsonData = {
-          source: source,
-          parseAsPng: isSlideDeck,
-        }
-        formData.append(
-          'source',
-          new Blob([JSON.stringify(jsonData)], { type: 'application/json' }),
-        )
-
-        const response = await fetch(`/chat/${courseId}/document`, {
-          method: 'POST',
-          body: formData,
-          headers: { HMS_API_TOKEN: userInfo.chat_token.token },
-        })
-
-        if (!response.ok) {
-          throw new Error(`Failed to upload ${file.name}`)
-        }
-
-        message.success(`${file.name} uploaded.`)
-        setCountProcessed((prev) => prev + 1)
-      } catch (e) {
-        message.error(`Failed to upload ${file.name}`)
-      }
-    }
-    getDocuments()
-  }
-
-  const handleDeleteDocument = async (record: any) => {
-    setLoading(true)
-    try {
-      const response = await fetch(
-        `/chat/${courseId}/${record.docId}/document`,
-        {
-          method: 'DELETE',
-          headers: {
-            'Content-Type': 'application/json',
-            HMS_API_TOKEN: userInfo.chat_token.token,
-          },
-        },
-      )
-
-      if (!response.ok) {
-        throw new Error(`Failed to upload ${File.name}`)
-      }
-
-      message.success('Document deleted.')
-      getDocuments()
-    } catch (e) {
-      message.error('Failed to delete document.')
-    } finally {
-      setLoading(false)
-    }
-  }
-
-  const addDocument = async () => {
-    setLoading(true)
-    try {
-      const formData = await form.validateFields()
-
-      if (documentType === 'URL') {
-        await addUrl(formData.url)
-      }
-
-      if (documentType === 'FILE') {
-        const files = fileList.map((file) => file.originFileObj)
-        await uploadFiles(files, formData.source)
-      }
-
-      setAddDocumentModalOpen(false)
-      setLoading(false)
-      setFileList([])
-      form.resetFields()
-      getDocuments()
-    } finally {
-      setLoading(false)
-    }
-  }
-
   return (
     <div className="m-auto my-5">
-      <Modal
-        title="Add a new document for your chatbot to use."
+      <AddChatbotDocumentModal
         open={addDocumentModalOpen}
-        onCancel={() => !loading && setAddDocumentModalOpen(false)}
-        footer={[
-          <Button key="ok" onClick={() => setAddDocumentModalOpen(false)}>
-            Cancel
-          </Button>,
-          <Button
-            key="ok"
-            type="primary"
-            onClick={addDocument}
-            disabled={loading}
-          >
-            Submit
-          </Button>,
-        ]}
-      >
-        <>
-          <div>
-            <p>
-              <strong>Accepted File Types:</strong> .docx, .pptx, .txt, .csv,
-              .pdf
-            </p>
-          </div>
-          <div className="mb-2 flex h-fit w-full items-center justify-center gap-2">
-            <div
-              className={`${
-                documentType == 'FILE'
-                  ? 'border-blue-300 text-black'
-                  : 'border-blue-100 text-gray-600'
-              } flex flex-grow cursor-pointer items-center justify-center gap-2 rounded-lg  border-2 p-5 hover:bg-blue-50`}
-              onClick={() => setDocumentType('FILE')}
-            >
-              <FileAddOutlined />
-              <p className="text-md font-semibold">Upload Files</p>
-            </div>
-            <div
-              className={`${
-                documentType == 'URL'
-                  ? 'border-blue-300 text-black'
-                  : 'border-blue-100 text-gray-600'
-              } flex flex-grow cursor-pointer items-center  justify-center gap-2 rounded-lg border-2  p-5 hover:bg-blue-50`}
-              onClick={() => setDocumentType('URL')}
-            >
-              <GithubOutlined />
-              <p className="text-md font-semibold">GitHub File</p>
-            </div>
-          </div>
-          <Form form={form}>
-            {documentType === 'URL' && (
-              <Form.Item
-                name="url"
-                rules={[
-                  {
-                    required: true,
-                    message: 'Please provide a document URL.',
-                  },
-                ]}
-              >
-                <Input placeholder="Enter URL for a pdf file..." />
-              </Form.Item>
-            )}
-            {documentType === 'FILE' && (
-              <>
-                <Form.Item
-                  name="files"
-                  rules={[
-                    {
-                      required: true,
-                      message: 'Please provide document files.',
-                    },
-                  ]}
-                >
-                  <Dragger {...props}>
-                    <p className="ant-upload-drag-icon">
-                      <InboxOutlined />
-                    </p>
-                    <p className="ant-upload-text">
-                      Click or drag file to this area to upload
-                    </p>
-                    <p className="ant-upload-hint">
-                      Support for a single or bulk upload. Strictly prohibited
-                      from uploading company data or other banned files.
-                    </p>
-                  </Dragger>
-                </Form.Item>
-                {loading && (
-                  <Progress
-                    percent={Math.round(
-                      (countProcessed / fileList.length) * 100,
-                    )}
-                  />
-                )}
-                <div className="flex items-center justify-between align-middle">
-                  <div className="flex">
-                    Parse document as slides{' '}
-                    <Tooltip
-                      title={
-                        'This will generate descriptions for images, which is particularly useful if the document is a slide deck or another image-heavy component'
-                      }
-                    >
-                      <QuestionCircleOutlined className="ml-2" />
-                    </Tooltip>
-                  </div>
-                  <Switch
-                    defaultChecked={isSlideDeck}
-                    className="mt-0 pt-0"
-                    disabled={false}
-                    onChange={(checked) => setIsSlideDeck(checked)}
-                  />
-                </div>
-                <p>
-                  Preview URL{' '}
-                  <Tooltip
-                    title={
-                      'This preview URL will be used to redirect your students to view this file. Make sure to include http header unless you want to redirect route on this site.'
-                    }
-                  >
-                    <QuestionCircleOutlined className="ml-2" />{' '}
-                  </Tooltip>
-                </p>
-                <Form.Item
-                  name="source"
-                  rules={[
-                    {
-                      required: true,
-                      message: 'Please provide a document preview URL.',
-                    },
-                    {
-                      type: 'url',
-                      message: 'Please enter a valid URL.',
-                    },
-                  ]}
-                >
-                  <Input placeholder="Enter document preview URL..." />
-                </Form.Item>
-              </>
-            )}
-          </Form>
-        </>
-      </Modal>
+        courseId={courseId}
+        onClose={() => setAddDocumentModalOpen(false)}
+        getDocuments={getDocuments}
+      />
       <div className="flex w-full items-center justify-between">
         <div className="">
           <h3 className="text-4xl font-bold text-gray-900">
